typecheck fixes
