+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/components/fleet_extensions/constants.ts
@@ -5,11 +5,11 @@
  * 2.0.
  */
 
-export const ASSET_POLICY_TEMPLATE = 'asset_inventory';
+import { CLOUDBEAT_AWS } from './aws_credentials_form/constants';
+import { CLOUDBEAT_AZURE } from './azure_credentials_form/constants';
+import { CLOUDBEAT_GCP } from './gcp_credentials_form/constants';
 
-export const CLOUDBEAT_AWS = 'cloudbeat/asset_inventory_aws';
-export const CLOUDBEAT_GCP = 'cloudbeat/asset_inventory_gcp';
-export const CLOUDBEAT_AZURE = 'cloudbeat/asset_inventory_azure';
+export const ASSET_POLICY_TEMPLATE = 'asset_inventory';
 
 export const SUPPORTED_POLICY_TEMPLATES = [ASSET_POLICY_TEMPLATE] as const;
 export const SUPPORTED_CLOUDBEAT_INPUTS = [CLOUDBEAT_AWS, CLOUDBEAT_GCP, CLOUDBEAT_AZURE] as const;
@@ -17,18 +17,13 @@ export const SUPPORTED_CLOUDBEAT_INPUTS = [CLOUDBEAT_AWS, CLOUDBEAT_GCP, CLOUDBE
 export const SINGLE_ACCOUNT = 'single-account';
 export const ORGANIZATION_ACCOUNT = 'organization-account';
 
-export const AWS_SINGLE_ACCOUNT = 'single-account';
-export const AWS_ORGANIZATION_ACCOUNT = 'organization-account';
-export const GCP_SINGLE_ACCOUNT = 'single-account';
-export const GCP_ORGANIZATION_ACCOUNT = 'organization-account';
-export const AZURE_SINGLE_ACCOUNT = 'single-account';
-export const AZURE_ORGANIZATION_ACCOUNT = 'organization-account';
-
 export const SUPPORTED_TEMPLATES_URL_FROM_PACKAGE_INFO_INPUT_VARS = {
   CLOUD_FORMATION: 'cloud_formation_template',
   CLOUD_FORMATION_CREDENTIALS: 'cloud_formation_credentials_template',
   ARM_TEMPLATE: 'arm_template_url',
   CLOUD_SHELL_URL: 'cloud_shell_url',
 };
+
+export const ASSET_NAMESPACE = 'default';
 
 export const TEMPLATE_URL_ACCOUNT_TYPE_ENV_VAR = 'ACCOUNT_TYPE';
