+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/components/fleet_extensions/gcp_credentials_form/constants.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const CLOUDBEAT_GCP = 'cloudbeat/asset_inventory_gcp';
+export const GCP_SINGLE_ACCOUNT = 'single-account';
+export const GCP_ORGANIZATION_ACCOUNT = 'organization-account';
