+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/components/fleet_extensions/gcp_credentials_form/types.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export type GcpCredentialsType = 'credentials-file' | 'credentials-json' | 'credentials-none';
+
+export type GcpCredentialsTypeFieldMap = {
+  [key in GcpCredentialsType]: string[];
+};
